requirements.txt
beautifulsoup4==4.8.2
requests==2.22.0
Jinja2==2.11.2
python-dotenv == 0.13.0
